Séparations des nodes et du scenegraph
